Implemented SWOS VM
Finally, nearly finished C++ converter for SWOS assembly code.
Main points:
- switched from C output to C++
- implemented alignment-safe memory access
- added some optimizations on the assembly code
- x64 version added!

+++ ida2asm/src/OutputItem/StackVariable.h
@@ -25,7 +25,7 @@ private:
     uint8_t m_offsetLength;
     uint8_t m_sizeLength;
     uint8_t m_size;
-    int8_t m_offset;
+    int16_t m_offset;
     // followed by: name string, offset string, size/struct string
 };
 #pragma pack(pop)
